calibration bug fix
This is preliminary try to fix the problem arising from in-definite
calib voltage points in calibration.
Use the peaks found to draw the calib graph.However,require all channels
have the same number of peaks.In case that some of the channels have
problem and have different number of peaks,the result is in-definite.
Another problem is that it does not consider the calib voltage
value.Only predefined voltage steps are used.
Need more flexible solution here.